requirements.txt
git+https://github.com/nerfstudio-project/gsplat@ec3e715f5733df90d804843c7246e725582df10c
pycolmap==3.10.0
viser
nerfview==0.0.2
imageio[ffmpeg]
numpy<2.0.0
scikit-learn
tqdm
torchmetrics[image]
opencv-python
tyro>=0.8.8
Pillow
tensorboard
tensorly
pyyaml
matplotlib
git+https://github.com/rahul-goel/fused-ssim@1272e21a282342e89537159e4bad508b19b34157
kornia
easydict
plotly
evo==1.11.0
plyfile

